refactor(admin): 抽取 admin_account / account_handler TK companion
Update/Create/Bulk 与 protocol-probe/models/list 迁出上游形文件；同步
protocol-routing SSOT 与 public-group aggregator 检查接受 companion 路径。
功能不变。no-web-impact
