+++ backend/internal/service/admin_account_tk_billing_probe.go
@@ -1,0 +1,12 @@
+package service
+
+import (
+	kiro "github.com/Wei-Shaw/sub2api/internal/pkg/kiro"
+)
+
+func normalizeAccountPriority(platform string, priority int) int {
+	if platform == PlatformKiro && priority <= 0 {
+		return kiro.DefaultKiroAccountPriority
+	}
+	return priority
+}
