+++ scripts/sentinels/grok.json
@@ -254,12 +254,18 @@
       "rationale": "Create-time validate+prime (resolveGrokTokenOnSave): pastes a grok refresh_token, does a live xAI refresh, and either primes the account (access_token/expires_at/token_endpoint) or rejects with the exact upstream reason. tkInputHasNonEmptyCredential gates the UpdateAccount re-validate so an unrelated grok edit isn't blocked by a transient xAI outage. Losing either lets dead grok accounts materialize, or fires an xAI call on every edit."
     },
     {
-      "path": "backend/internal/service/admin_account.go",
+      "path": "backend/internal/service/admin_account_tk_create.go",
+      "rationale": "The Grok validate+prime hook must remain wired at both save boundaries: CreateAccount always validates a newly pasted token before conservative official-profile protocol seeding and persistence, while UpdateAccount revalidates only when that request explicitly supplies a non-empty refresh_token. The distinct surrounding blocks prevent a surviving create call from masking deletion of the update call, prevent protocol seeding from persisting an unvalidated account, and prevent unrelated Grok edits from acquiring a live xAI dependency. (create boundary companion)",
       "must_contain": [
-        "if err := resolveGrokTokenOnSave(ctx, account); err != nil {\n\t\treturn nil, err\n\t}\n\tSeedOfficialSupportedProtocols(account)\n\tif err := s.accountRepo.Create",
+        "if err := resolveGrokTokenOnSave(ctx, account); err != nil {\n\t\treturn nil, err\n\t}\n\tSeedOfficialSupportedProtocols(account)\n\tif err := s.accountRepo.Create"
+      ]
+    },
+    {
+      "path": "backend/internal/service/admin_account_tk_update.go",
+      "rationale": "The Grok validate+prime hook must remain wired at both save boundaries: CreateAccount always validates a newly pasted token before conservative official-profile protocol seeding and persistence, while UpdateAccount revalidates only when that request explicitly supplies a non-empty refresh_token. The distinct surrounding blocks prevent a surviving create call from masking deletion of the update call, prevent protocol seeding from persisting an unvalidated account, and prevent unrelated Grok edits from acquiring a live xAI dependency. (update boundary companion)",
+      "must_contain": [
         "if account.Platform == PlatformGrok &&\n\t\ttkInputHasNonEmptyCredential(input.Credentials, \"refresh_token\") {\n\t\tif err := resolveGrokTokenOnSave(ctx, account); err != nil {"
-      ],
-      "rationale": "The Grok validate+prime hook must remain wired at both save boundaries: CreateAccount always validates a newly pasted token before conservative official-profile protocol seeding and persistence, while UpdateAccount revalidates only when that request explicitly supplies a non-empty refresh_token. The distinct surrounding blocks prevent a surviving create call from masking deletion of the update call, prevent protocol seeding from persisting an unvalidated account, and prevent unrelated Grok edits from acquiring a live xAI dependency."
+      ]
     },
     {
       "path": "backend/internal/service/account_test_service.go",
