+++ scripts/sentinels/newapi.json
@@ -137,12 +137,18 @@
       "rationale": "TK companion invoked from admin_service.go CreateAccount/UpdateAccount that pins newapi/Moonshot accounts to the regional base_url their API key actually accepts. Without it, admins silently save accounts with the wrong region (api.moonshot.cn vs .ai), and the relay hot path 401s every request because it deliberately does not do per-request region fallback. Build-break alone is insufficient: an upstream merge that simultaneously removes the call sites would silently regress this without compile failure."
     },
     {
-      "path": "backend/internal/service/admin_account.go",
+      "path": "backend/internal/service/admin_account_tk_create.go",
+      "rationale": "CreateAccount and UpdateAccount are separate persistence boundaries and must both invoke the Moonshot regional resolver. The two context-specific anchors distinguish the create call (followed by unconditional Grok create validation) from the update call (followed by conditional Grok refresh validation), so one surviving call cannot mask deletion of the other. (create boundary companion)",
       "must_contain": [
-        "if err := resolveNewAPIMoonshotBaseURLOnSave(ctx, account); err != nil {\n\t\treturn nil, err\n\t}\n\tif err := resolveGrokTokenOnSave",
-        "if err := resolveNewAPIMoonshotBaseURLOnSave(ctx, account); err != nil {\n\t\treturn nil, err\n\t}\n\tif account.Platform == PlatformGrok"
-      ],
-      "rationale": "CreateAccount and UpdateAccount are separate persistence boundaries and must both invoke the Moonshot regional resolver. The two context-specific anchors distinguish the create call (followed by unconditional Grok create validation) from the update call (followed by conditional Grok refresh validation), so one surviving call cannot mask deletion of the other."
+        "if err := resolveNewAPIMoonshotBaseURLOnSave(ctx, account); err != nil {\n\t\treturn nil, err\n\t}\n\tif err := resolveGrokTokenOnSave"
+      ]
+    },
+    {
+      "path": "backend/internal/service/admin_account_tk_update.go",
+      "rationale": "CreateAccount and UpdateAccount are separate persistence boundaries and must both invoke the Moonshot regional resolver. The two context-specific anchors distinguish the create call (followed by unconditional Grok create validation) from the update call (followed by conditional Grok refresh validation), so one surviving call cannot mask deletion of the other. (update boundary companion)",
+      "must_contain": [
+        "if err := resolveNewAPIMoonshotBaseURLOnSave(ctx, account); err != nil {\n\t\treturn err\n\t}\n\tif account.Platform == PlatformGrok"
+      ]
     },
     {
       "path": "backend/internal/integration/newapi/channel_types.go",
